+++ src/main/java/de/crfa/app/resource/domain/ReleaseDto.java
@@ -14,7 +14,7 @@ public class ReleaseDto {
 
     String releaseName;
 
-    int releaseNumber;
+    float releaseNumber;
 
     String description;
